recipes: anonymous functions with priorities
The execution order of anonymous functions inside a recipe is
deterministic (executed in the order in which the functions get
parsed), but class authors cannot know in which order classes get
inherited and thus cannot know whether the class functions run before
or after functions in other classes. That is a problem for
rm_work.bbclass, which must add its own task after all other
classes have added theirs.

As an extension of the current syntax, now any function whose name
*starts* with "__anonymous" is considered an anonymous function;
previously, anything with *exactly* that name was an anonymous
function.

In contrast to those which are named just "__anonymous" or have no
name at all, the "__anonymous_<something>" name must be globally
unique and can be used to set a "__anonprio" variable flag for the
function.

The default priority is 100. Functions with higher values run later.

For example, do_rm_work.bbclass can use this to ensure that it runs
later than other anonymous functions:

python __anonymous_rm_work() {
    bb.build.addtask(....)
}
__anonymous_rm_work[priority] = "1000"

The priorities influence parsing results and thus need to be included
in the base hash. That part was necessary to trigger re-parsing when
modifying the priority in rm_work.bbclass, which gets inherited
globally via INHERIT.

Related-to: [YOCTO #10584]

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>