rm_work.bbclass: add do_rm_work after other classes are done
This uses the new bitbake support for prioritizing anonymous functions
to ensure that rm_work.bbclass gets to see a full set of existing
tasks when adding its own one. This is relevant, for example, for
do_analyseimage in meta-isa-fw's isafw.bbclass.

It also fixes a problem with finding the do_package_write_rpm/ipk/deb
dependency: those are not in the 'deps' variable. They can be found
via the do_build[recrdeptask] variable flag.

With both fixed it is no longer necessary to hard-code certain known
other tasks.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>